Mapbox integration

+++ app/src/main/java/com/deepred/subworld/utils/IMarkersListener.java
@@ -2,7 +2,8 @@ package com.deepred.subworld.utils;
 
 import android.location.Location;
 
-import com.google.android.gms.maps.model.LatLng;
+import com.mapbox.mapboxsdk.geometry.LatLng;
+//import com.google.android.gms.maps.model.LatLng;
 
 /**
  *
@@ -12,5 +13,5 @@ public interface IMarkersListener {
     public void updateMyMarker(Location loc);
     public void removeMarker(String uid);
     public void providerChanged(boolean isGps);
-    public void setZoom(float zoom);
+    public void setZoom(double zoom);
 }
